rename of SoftJointLimiter to JointSoftLimiter and some minor changes from manuel
